import as per GJS ESM documentation
