+++ @types/gnome-shell/extensions/misc/util.d.ts
@@ -1,0 +1,2 @@
+export declare function spawn(argv: string[]): void;
+export declare function lerp(start: number, end: number, progress: number): number;
