+++ @types/gnome-shell/extensions/ui/layout.d.ts
@@ -1,0 +1,5 @@
+import Clutter from '../../../clutter12';
+
+declare class MonitorConstraint extends Clutter.Constraint {
+    constructor(params: Partial<{ primary: boolean, index: number }>);
+}
