+++ extension/src/altTab.ts
@@ -2,12 +2,12 @@ import Clutter from '@gi-types/clutter';
 import GLib from '@gi-types/glib2';
 import Shell from '@gi-types/shell';
 import St from '@gi-types/st';
-import { imports } from 'gnome-shell';
+import Main from 'resource:///org/gnome/Shell/Extensions/js/extensions/ui/main';
+import { WindowSwitcherPopup } from 'resource:///org/gnome/Shell/Extensions/js/extensions/ui/altTab';
+import { __shell_private_types } from 'resource:///org/gnome/Shell/Extensions/js/extensions/ui/swipeTracker';
+
 import { AltTabConstants, ExtSettings } from '../constants';
 import { TouchpadSwipeGesture } from './swipeTracker';
-
-const Main = imports.ui.main;
-const { WindowSwitcherPopup } = imports.ui.altTab;
 
 let dummyWinCount = AltTabConstants.DUMMY_WIN_COUNT;
