+++ extension/src/pinchGestures/closeWindow.ts
@@ -3,16 +3,15 @@ import Meta from '@gi-types/meta';
 import Shell from '@gi-types/shell';
 import St from '@gi-types/st';
 
-import { global, imports } from 'gnome-shell';
+import Main from 'resource:///org/gnome/Shell/Extensions/js/extensions/ui/main';
+import { global } from 'resource:///org/gnome/Shell/Extensions/js/extensions/ui/swipeTracker';
+import * as Util from 'resource:///org/gnome/Shell/Extensions/js/extensions/misc/util';
 
 import { PinchGestureType } from '../../common/settings';
 import { WIGET_SHOWING_DURATION } from '../../constants';
 import { TouchpadPinchGesture } from '../trackers/pinchTracker';
 import { easeActor } from '../utils/environment';
 import { getVirtualKeyboard, IVirtualKeyboard } from '../utils/keyboard';
-
-const Main = imports.ui.main;
-const Util = imports.misc.util;
 
 const END_OPACITY = 0;
 const END_SCALE = 0.5;
