+++ extension/src/swipeTracker.ts
@@ -2,13 +2,12 @@ import Clutter from '@gi-types/clutter';
 import GObject from '@gi-types/gobject2';
 import Meta from '@gi-types/meta';
 import Shell from '@gi-types/shell';
-import { CustomEventType, global, imports } from 'gnome-shell';
+import { SwipeTracker, CustomEventType, global } from 'resource:///org/gnome/Shell/Extensions/js/extensions/ui/swipeTracker';
+import * as Main from 'resource:///org/gnome/Shell/Extensions/js/extensions/ui/main';
 import { registerClass } from '../common/utils/gobject';
 import { TouchpadConstants } from '../constants';
 import * as DBusUtils from './utils/dbus';
 
-const Main = imports.ui.main;
-const { SwipeTracker } = imports.ui.swipeTracker;
 
 // define enum
 enum TouchpadState {
