+++ package.json
@@ -7,6 +7,7 @@
   "scripts": {
     "generate-bindings": "sh scripts/generate-gi-bindings.sh",
     "test": "echo \"Error: no test specified\" && exit 1",
+    "build:transpile-from-config": "tsc --build tsconfig.json",
     "build:transpile": "tsc scripts/transpile.ts --outDir build --allowSyntheticDefaultImports --sourceMap --target esnext --module esnext --moduleResolution node",
     "lint:extension": "eslint extension scripts tests",
     "lint:package": "eslint build --fix",
@@ -37,10 +38,10 @@
     "@types/glob": "^7.2.0",
     "@types/node": "^17.0.23",
     "@types/yargs": "^17.0.10",
-    "@typescript-eslint/eslint-plugin": "^5.18.0",
-    "@typescript-eslint/parser": "^5.18.0",
-    "typescript": "^4.3.5",
-    "eslint": "^8.12.0"
+    "@typescript-eslint/eslint-plugin": "^6.7.5",
+    "@typescript-eslint/parser": "^6.7.5",
+    "typescript": "^5.2.2",
+    "eslint": "^8.51.0"
   },
   "dependencies": {
     "fs": "0.0.1-security",
